dom/imptests/webapps/DOMCore/tests/submissions/Ms2ger/test_CharacterData-appendData.html

<!DOCTYPE html>
<title>CharacterData.appendData</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
function testNode(node) {
  test(function() {
    assert_equals(node.data, "test")
    node.appendData("test")
    assert_equals(node.data, "testtest")
  })
}
test(function() {
  testNode(document.createTextNode("test"))
  testNode(document.createComment("test"))
})
</script>
